+++ src/main/java/com/example/ComUniShare/controllers/UserController.java
@@ -3,11 +3,14 @@ package com.example.ComUniShare.controllers;
 import com.example.ComUniShare.domain.user.User;
 import com.example.ComUniShare.domain.user.UserResponseDTO;
 import com.example.ComUniShare.services.user.UserService;
+import org.hibernate.annotations.NotFound;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @RestController()
@@ -17,6 +20,7 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    //ok
     @GetMapping("/{userId}")
     public ResponseEntity<UserResponseDTO> getUserById(@PathVariable String userId) {
         User user = userService.findById(userId);
@@ -29,6 +33,7 @@ public class UserController {
         }
     }
 
+    //ok
     @GetMapping("/login/{login}")
     public ResponseEntity<UserResponseDTO> getUserByName(@PathVariable String login) {
         User user = userService.findUserByLogin(login);
@@ -41,6 +46,7 @@ public class UserController {
         }
     }
 
+    //ok
     @GetMapping("/mydata")
     public ResponseEntity<UserResponseDTO> getMyData() {
         User user = userService.myData();
@@ -53,6 +59,7 @@ public class UserController {
         }
     }
 
+    //ok
     @GetMapping
     public ResponseEntity<List<UserResponseDTO>> getAllUsers() {
         List<User> users = userService.findAllUsers();
@@ -60,15 +67,19 @@ public class UserController {
         return ResponseEntity.ok(userDTOs);
     }
 
+    //ok
     @PutMapping("/{userId}")
-    public ResponseEntity<String> updateUser(@PathVariable String userId, @RequestBody User user) {
-        user.setId(userId);
-        userService.updateUser(user);
-        return ResponseEntity.ok("Usuário atualizado com sucesso!");
+    public ResponseEntity<String> updateUser(@PathVariable String userId, @RequestBody Map<String, Object> updates) {
+        return userService.updateUserAdapted(userId, updates);
     }
 
+    //ok
     @DeleteMapping("/{userId}")
     public ResponseEntity<String> deleteUser(@PathVariable String userId) {
+        User user = userService.findById(userId);
+        if (user == null) {
+            return new ResponseEntity<>("Usuário não encontrado com ID: " + userId, HttpStatus.NOT_FOUND);
+        }
         userService.deleteUser(userId);
         return ResponseEntity.ok("Usuário excluído com sucesso!");
     }
